+++ api/src/routes/categories.route.js
